All helper functions and classes in one file.
